+++ src/core/api/handlers/sdk.js
@@ -2,7 +2,8 @@
  * @module SDK
  */
 
-import { logger } from '../../_helpers_/optimizelyHelper.js';
+import { logger } from '../../../utils/optimizelyHelper.js';
+import defaultSettings from '../../../utils/config/defaultSettings.js';
 
 /**
  * Fetches and updates the Optimizely JavaScript SDK based on the provided URL.
